Fix bundler 2.0 development setup
Allow bundler 2.0 as development dependency.

Bundler 2.0 is not compatible with all Ruby or rubygems versions used in
builds. Disable Ruby 2.2 in build matrix and skip bundle install step.

Runing tests on Rails 4.2 and Ruby 2.6.0 fails. That is expected as Rails 4.2.0 is not maintained. Run tests only against Rails 4.2.0 only on Ruby 2.3.8.